requirements.txt
azure-ai-textanalytics==5.2.0b4
azure-common==1.1.28
azure-core==1.24.1
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.0
python-dotenv==0.20.0
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
